+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## 0.8.3 (June 4, 2018)
+
+- Updated parameter value checks in Deploy-HyperVVagrantBoxManually function
+- Removed Message about.\nuget.exe path on Module import
+
 ## 0.8.2 (June 1, 2018)
 
 - Created